# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Contacts API",
    "version": "1.0.0",
    "description": "API for managing contacts"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://cse-341-project01-isr0.onrender.com",
      "description": "Production server on Render"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "summary": "Returns the list of all users",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "The list of the users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The user was successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get the user by id",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The user id"
          }
        ],
        "responses": {
          "200": {
            "description": "The user description by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "summary": "Update the user by the id",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The user id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user was updated"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "summary": "Remove the user by id",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The user id"
          }
        ],
        "responses": {
          "204": {
            "description": "The user was deleted successfully"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "firstName",
          "lastName",
          "email",
          "favoriteColor",
          "birthday"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "The auto-generated id of the user"
          },
          "firstName": {
            "type": "string",
            "description": "User's first name"
          },
          "lastName": {
            "type": "string",
            "description": "User's last name"
          },
          "email": {
            "type": "string",
            "description": "User's email address"
          },
          "favoriteColor": {
            "type": "string",
            "description": "User's favorite color"
          },
          "birthday": {
            "type": "string",
            "format": "date",
            "description": "User's birthday"
          }
        },
        "example": {
          "_id": "60df1f1b2f4d4e0015d39b25",
          "firstName": "Chinedu",
          "lastName": "Okoro",
          "email": "chinedu.okoro@example.com",
          "favoriteColor": "Purple",
          "birthday": "1988-03-12"
        }
      }
    }
  },
  "tags": [
    {
      "name": "Users",
      "description": "The users managing API"
    }
  ]
}